project/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.0
supabase==2.3.4
openai==1.12.0
sqlalchemy==2.0.25
numpy==1.26.3
scipy==1.12.0